+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -6,7 +6,10 @@ import { authController } from "~/modules/auth/auth.js";
 import { filesController } from "~/modules/files/files.js";
 import { notificationController } from "~/modules/notifications/notifications.js";
 import { pointsOfInterestController } from "~/modules/points-of-interest/points-of-interest.js";
+import { reviewController } from "~/modules/reviews/reviews.js";
 import { routeCategoryController } from "~/modules/route-categories/route-categories.js";
+import { routesController } from "~/modules/routes/routes.js";
+import { userFollowsController } from "~/modules/user-follows/user-follows.js";
 import { userController } from "~/modules/users/users.js";
 
 import { BaseServerApplicationApi } from "./base-server-application-api.js";
@@ -29,10 +32,13 @@ const apiV1 = new BaseServerApplicationApi(
 	},
 	...authController.routes,
 	...filesController.routes,
-	...routeCategoryController.routes,
-	...userController.routes,
 	...notificationController.routes,
 	...pointsOfInterestController.routes,
+	...reviewController.routes,
+	...routeCategoryController.routes,
+	...routesController.routes,
+	...userController.routes,
+	...userFollowsController.routes,
 );
 
 const serverApplication = new BaseServerApplication({