+++ apps/frontend/src/libs/types/navigation-item.type.ts
@@ -4,6 +4,7 @@ import { type IconName, type ValueOf } from "~/libs/types/types.js";
 type NavigationItem = {
 	href: ValueOf<typeof AppRoute>;
 	icon: IconName;
+	isLabelHidden?: boolean;
 	label: string;
 };