+++ apps/frontend/src/modules/files/libs/types/types.ts
@@ -1,5 +1,4 @@
 export {
 	type FileContentType,
-	type FileUploadUrlRequestDto,
 	type FileUploadUrlResponseDto,
 } from "@smartscapes/shared";