+++ packages/shared/src/index.ts
@@ -52,11 +52,31 @@ export {
 	pointsOfInterestSearchQueryValidationSchema,
 } from "./modules/points-of-interest/points-of-interest.js";
 export {
+	type ReviewCreatePayload,
+	reviewCreateValidationSchema,
+	type ReviewGetByIdResponseDto,
+	type ReviewRequestDto,
+} from "./modules/reviews/reviews.js";
+export {
 	RouteCategoriesApiPath,
 	RouteCategoryExceptionMessage,
 	type RouteCategoryGetAllItemResponseDto,
 	type RouteCategoryRequestDto,
 } from "./modules/route-categories/route-categories.js";
+export {
+	routesCreateValidationSchema,
+	type RoutesRequestCreateDto,
+	type RoutesRequestPatchDto,
+	type RoutesResponseDto,
+	routesUpdateValidationSchema,
+} from "./modules/routes/routes.js";
+export {
+	UserFollowsApiPath,
+	UserFollowsExceptionMessage,
+	type UserFollowsParametersDto,
+	type UserFollowsRequestDto,
+	type UserUnfollowsParametersDto,
+} from "./modules/user-follows/user-follows.js";
 export {
 	type UserAuthResponseDto,
 	UserExceptionMessage,