+++ packages/shared/src/libs/enums/api-path.enum.ts
@@ -3,7 +3,9 @@ const APIPath = {
 	FILES: "/files",
 	NOTIFICATIONS: "/notifications",
 	POINTS_OF_INTEREST: "/points-of-interest",
+	REVIEWS: "/reviews",
 	ROUTE_CATEGORIES: "/route-categories",
+	ROUTES: "/routes",
 	USERS: "/users",
 } as const;