+++ packages/shared/src/libs/enums/content-type.enum.ts
@@ -1,13 +1,5 @@
 const ContentType = {
-	BINARY: "application/octet-stream",
-	GIF: "image/gif",
-	JPEG: "image/jpeg",
-	JPG: "image/jpg",
 	JSON: "application/json",
-	PDF: "application/pdf",
-	PNG: "image/png",
-	WEBP: "image/webp",
-	ZIP: "application/zip",
 } as const;
 
 export { ContentType };