+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Changelog
 
+## [1.14.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/backend-v1.13.2...backend-v1.14.0) (2025-08-10)
+
+
+### Features
+
+* **backend/frontend:** add logout functionality ss-182 ([#249](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/249)) ([e3a03fb](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/e3a03fbdc47c9d509d5aef7eb6a77866e122338f))
+* **backend/shared:** add followers module ss-72 ([#136](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/136)) ([3b63b29](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/3b63b2952a6b4e62d7e46b91fe2825f2c307c509))
+
+
+### Bug Fixes
+
+* **backend:** format swagger docs ss-285 ([#286](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/286)) ([00f42c0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/00f42c021b091dc18f181ee246f4531d99fff69b))
+* **backend:** update notification swagger response dto ss-279 ([#289](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/289)) ([c111a97](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/c111a97f3b1294de175f9160a75b10b07685c52f))
+
+## [1.13.2](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/backend-v1.13.1...backend-v1.13.2) (2025-08-07)
+
+
+### Bug Fixes
+
+* **backend:** fix poi location schema in swagger documentation ss-258 ([#259](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/259)) ([323d2cf](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/323d2cfb51ba8c817bff46f00b9d079efffacca7))
+
+## [1.13.1](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/backend-v1.13.0...backend-v1.13.1) (2025-08-07)
+
+
+### Bug Fixes
+
+* **backend:** update reviews module swagger docs ss-251 ([#253](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/253)) ([4753c0e](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/4753c0e243a7d4d8b39ae271c911b0bfe5995d51))
+
+## [1.13.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/backend-v1.12.0...backend-v1.13.0) (2025-08-07)
+
+
+### Features
+
+* **backend/shared:** add reviews module ss-75 ([#147](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/147)) ([435363e](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/435363e7ee382997fab923744e484ecc0efde238))
+
+## [1.12.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/backend-v1.11.0...backend-v1.12.0) (2025-08-07)
+
+
+### Features
+
+* **backend/shared:** add routes module ss-69 ([#104](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/104)) ([b419013](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/b419013fd741b365045a9e6f3d49c30a1885a6b7))
+* **frontend:** update buttons in header ss-186 ([#210](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/210)) ([e4f8d7f](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/e4f8d7ffe91ea72f18bc8932a086e4c7e1194f42))
+
 ## [1.11.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/backend-v1.10.0...backend-v1.11.0) (2025-08-06)